fix getbyid function with check in database

+++ src/services/contactsService.js
@@ -1,4 +1,5 @@
 const { Contact } = require('../db/contactsDB');
+var mongoose = require('mongoose');
 
 const listContactsService = async () => {
   const data = await Contact.find();
@@ -7,10 +8,16 @@ const listContactsService = async () => {
 
 const getContactByIdService = async (id) => {
   const data = await Contact.findById(id)
-  return {status: 200, message: data}
+  return !data ? {status: 404, message: {"message": "Not found"}} : {status: 200, message: data}
 }
 
 const removeContactService = async (id) => {
+  if(!id){
+    return {status: 400, message: {"message": "Enter user id!"}}
+  }
+  if(!await getContactByIdService(id)){
+    return {status: 404, message: {"message": "Not found!"}}
+  }
   await Contact.findByIdAndRemove(id)
   return {status: 200, message: {"message": "contact deleted"}}
 }
@@ -21,6 +28,9 @@ const addContactService = async (body) => {
 }
 
 const updateContactService = async (id, body) => {
+  if(!id){
+    return {status: 404, message: {"message": "Not found!"}}
+  }
   const data = await Contact.findByIdAndUpdate(id, {...body})
   return {status: 200, message: data}
 }
